requirements.txt
matplotlib==3.7.1
PyQt5==5.15.9
Pillow==9.5.0
